+++ guicey-server-pages/src/test/groovy/ru/vyarus/guicey/gsp/views/ViewsSubMappingTest.groovy
@@ -9,8 +9,12 @@ import ru.vyarus.dropwizard.guice.test.spock.ConfigOverride
 import ru.vyarus.dropwizard.guice.test.spock.UseDropwizardApp
 import ru.vyarus.guicey.gsp.AbstractTest
 import ru.vyarus.guicey.gsp.ServerPagesBundle
+import ru.vyarus.guicey.gsp.info.GspInfoService
+import ru.vyarus.guicey.gsp.info.model.GspApp
 import ru.vyarus.guicey.gsp.support.app.OverridableTemplateResource
 import ru.vyarus.guicey.gsp.support.app.SubTemplateResource
+
+import javax.inject.Inject
 
 /**
  * @author Vyacheslav Rusakov
@@ -20,6 +24,9 @@ import ru.vyarus.guicey.gsp.support.app.SubTemplateResource
         @ConfigOverride(key = "server.rootPath", value = "/rest/*")
 ])
 class ViewsSubMappingTest extends AbstractTest {
+
+    @Inject
+    GspInfoService info
 
     def "Check app mapped"() {
 
@@ -32,6 +39,21 @@ class ViewsSubMappingTest extends AbstractTest {
         res = getHtml("/app/sub/sample")
         then: "index page"
         res.contains("page: /app/sub/sample")
+
+        when: "get info"
+        GspApp app = info.getApplication("app")
+        then: "info correct"
+        with(app.getViews()) {
+            size() == 2
+            it[""] == "app/"
+            it["sub/"] == "sub/"
+        }
+        with(app.getAssets()){
+            size() == 1
+            it.get("") == ["app/"]
+        }
+        app.indexFile == "index.html"
+
     }
 
     static class App extends Application<Configuration> {